Refactor food photo to standalone component
